created separate module for getting server routes
